backend/requirements_latest.txt
fastapi
uvicorn
requests
azure-ai-documentintelligence
pydantic
pyyaml
pydantic-settings
python-dotenv
SQLAlchemy
pytz
tiktoken
openai
alembic
markitdown
aiofiles
playwright
asyncpg
gunicorn
ruff
python-multipart
pydantic-core
pydantic-settings
aiokafka